requirements.txt
Flask==1.0.2
line-bot-sdk==1.8.0
beautifulsoup4==4.7.1
urllib3==1.24.1
soupsieve==1.6.1
psycopg2